change admin tabs to use react-router
