+++ src/components/Routes.jsx
@@ -27,17 +27,20 @@ const Routes = () => {
     }
   );
   return (
-    <React.Fragment>
+    <>
       <Route path="/" exact component={LandingScreen} />
       <Route path="/releases" exact component={ReleasesScreen} />
-      <Route path="/admin/" exact component={AdminScreen} />
+      <Route path="/admin/">
+        <AdminScreen />
+      </Route>
+
       <Route
-        path="/admin/releasenotes/edit/:id"
+        path="/releasenotes/edit/:id"
         exact
         render={props => <ReleaseNoteEditorScreen {...props} />}
       />
       <Route
-        path="/admin/releasenotes/create"
+        path="/releasenotes/create"
         exact
         component={ReleaseNoteEditorScreen}
       />
@@ -47,17 +50,9 @@ const Routes = () => {
         render={props => <ReleaseScreen {...props} />}
       />
       <Route path="/login/" exact component={LoginScreen} />
-      <Route
-        path="/admin/releases/create"
-        exact
-        component={ReleaseEditorScreen}
-      />
-      <Route
-        path="/admin/releases/edit/:id"
-        exact
-        component={ReleaseEditorScreen}
-      />
-    </React.Fragment>
+      <Route path="/releases/create" exact component={ReleaseEditorScreen} />
+      <Route path="/releases/edit/:id" exact component={ReleaseEditorScreen} />
+    </>
   );
 };
